pyproject.toml
[project]
name = "mobilesam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.8",
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
    "lightning-app>=2.3.1",
    "lvis>=0.5.3",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "onnx>=1.19.1",
    "onnxruntime>=1.23.1",
    "pillow>=11.3.0",
    "pycocotools>=2.0.10",
    "rich>=13.9.4",
    "ruff>=0.14.6",
    "samexporter>=0.3.0",
    "scikit-image>=0.25.2",
    "segment-anything>=1.0",
    "segmentation-models-pytorch>=0.5.0",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torch-pruning>=1.6.1",
    "wandb>=0.21.4",
]

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-gpu" }
torchvision = { index = "pytorch-gpu" }
torchaudio = { index = "pytorch-gpu" }

[tool.ruff]
line-length = 100
target-version = "py310"

exclude = [
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
    "N",
    "NPY",
    "RUF"
]

ignore = [
    "E501",
    "F401",
]

[tool.ruff.lint.isort]
known-first-party = ["mobile_sam"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
